Changed how Futures are bound to Task completion
