+++ src/core/scala/app/com/linkedin/playparseq/s/PlayParSeq.scala
@@ -16,7 +16,9 @@ import com.linkedin.playparseq.s.stores.ParSeqTaskStore
 import com.linkedin.playparseq.utils.PlayParSeqHelper
 import javax.inject.{Inject, Singleton}
 import play.api.mvc.RequestHeader
+
 import scala.concurrent.Future
+import scala.concurrent.Promise
 
 
 /**
@@ -85,13 +87,14 @@ class PlayParSeqImpl @Inject()(engine: Engine, parSeqTaskStore: ParSeqTaskStore)
   }
 
   override def runTask[T](task: Task[T])(implicit requestHeader: RequestHeader): Future[T] = {
+    val scalaPromise: Promise[T] = Promise[T]()
     // Bind a Future to the ParSeq Task
-    val future: Future[T] = bindTaskToFuture(task)
+    val boundTask: Task[T] = bindTaskToFuture(task, scalaPromise)
     // Put the ParSeq Task into store
-    parSeqTaskStore.put(task)
+    parSeqTaskStore.put(boundTask)
     // Run the ParSeq Task
-    engine.run(task)
+    engine.run(boundTask)
     // Return the Future
-    future
+    scalaPromise.future
   }
 }
